+++ bridge-project/src/page/Card/CardReadOnly.tsx
