+++ bridge-project/src/page/Card/index.tsx
