+++ bridge-project/src/page/ver1/Detail.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import Tab from './Tab';
-import { TabModel } from './model/tabModel';
-import { data, data2, total } from './assets/dummy.ts';
-import { Info } from './model/info';
-import CardReadOnly from './Card/CardReadOnly.tsx';
-import Header from './Header/index.tsx';
+import Tab from '../FamPage/Tab.tsx';
+import { TabModel } from '../../model/tabModel.ts';
+import { data, data2, total } from '../../assets/dummy.ts';
+import { Info } from '../../model/info.ts';
+import CardReadOnly from '../Card/CardReadOnly.tsx';
+import Header from '../Header/index.tsx';
 
 const Detail: React.FC = () => {
   const curDate: string = '2024-07-17';
